br-ext: stm32mp1: sdcard part3 used as a EFI system partition
Changes partition 3 of the generated SDcard to be a EFI system partition
instead of a generic data partition. This is needed for U-Boot/EFI
to store EFI variables. We use FAT format since U-Boot is currently
bugged (at least up to v2022.07) when ESP is Ext4 formatted due to
"ubootefi.var" file not being an absolute path.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>